# disable precompilation for now